lab10/pyproject.toml
[project]
name = "lab8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "requests>=2.32.5",
]
